tree-walk: don't move to start of bounds
This is likely breaking something, since I don't have any tests to speak of.  But I can't remember why I did this, and it breaks things in an obvious way.  Anyway, for the moment it fixes something, so something on hand that I care about now is better than something that I can't remember, I guess.

More specifically, this change fixes transposition for treesitter objects, which I want.